Play around with caching shared effects and rayon
So far, just filtering in parallel is the fastest by far. Even when giving rayon only 4 threads it still beats sequential filtering + hashmap cache.

For some reason, the moka cache is the slowest of all. When using it in parallel with rayon it's about a second slower than sequential filtering + hashmap cache. I think I might be using it wrong somehow :confused:

Anyway, for now I'll stick with rayon and no cache :rocket: