feat: smarter filtering of potions with 3 ingredients
Ensures that each ingredient in a suggested 3-ingredient potion
contributes at least one new effect to the mixture.

Performance wise, the filtering is more complex and therefore slower.
However, it does reduce the amount of Potion instances created and the
time saved there is about equal to the time lost to the new filter.
In other words, wall time remains about the same but memory usage is
(slightly) better with this change.