feat: remove unused plugins from load order in GameData
Makes for a cleaner data export.

There seems to be a bug in the implementation where it does not remove
all the unused entries from the load order in one call. I haven't
noticed this cause any problems so far so I'll call this "good enough"
and look into it later.